use size_t instead of long

+++ include/nmpm.h
@@ -15,7 +15,7 @@ extern "C" {
 
 #if HAVE_CUDA
 #include"compute_gpu.cuh"
-int run_overlap_benchmark_gpu(int rank, int size, int dim, int compToPureCommRatio, long min_bytes,long max_bytes, int do_progress, int enable_thread, int compute_bound, memory_mode_t memory_mode);
+int run_overlap_benchmark_gpu(int rank, int size, int dim, int compToPureCommRatio, size_t min_bytes,size_t max_bytes, int do_progress, int enable_thread, int compute_bound, memory_mode_t memory_mode);
 #endif
 
 
@@ -45,7 +45,7 @@ int run_overlap_benchmark_gpu(int rank, int size, int dim, int compToPureCommRat
 #endif
 
 void usage(char *prog_name);
-int run_overlap_benchmark(int rank,int size, int dim, int compToPureCommRatio, long min_bytes, long max_bytes, int compute_bound,memory_mode_t memory_mode, int do_progress);
+int run_overlap_benchmark(int rank,int size, int dim, int compToPureCommRatio, size_t min_bytes, size_t max_bytes, int compute_bound,memory_mode_t memory_mode, int do_progress);
 
 #ifdef __cplusplus
 }
