+++ src/nmpm.c
@@ -139,7 +139,7 @@ static void post_sendrecv(int left,int right, int front, int back, int bottom, i
 
 }
 
-int run_overlap_benchmark(int rank, int size, int dim, int compToPureCommRatio, long min_bytes,long max_bytes, int compute_bound,memory_mode_t memory_mode, int do_progress) { 
+int run_overlap_benchmark(int rank, int size, int dim, int compToPureCommRatio, size_t min_bytes,size_t max_bytes, int compute_bound,memory_mode_t memory_mode, int do_progress) { 
 	int iter;
     double t_pure_total=0.0, t_comp_total=0.0, t_ovrl_total=0.0;
     double overlap=0.0;
@@ -284,7 +284,7 @@ int run_overlap_benchmark(int rank, int size, int dim, int compToPureCommRatio, 
 
 #if HAVE_CUDA
 
-int run_overlap_benchmark_gpu(int rank, int size, int dim, int compToPureCommRatio, long min_bytes, long max_bytes,int do_progress, int enable_thread, int compute_bound, memory_mode_t memory_mode) {
+int run_overlap_benchmark_gpu(int rank, int size, int dim, int compToPureCommRatio, size_t min_bytes, size_t max_bytes,int do_progress, int enable_thread, int compute_bound, memory_mode_t memory_mode) {
 	int iter;
 	int gpu_inner_iters;
 	int hostname_len;
